move NodeConfig to node-core
