+++ crates/node-core/src/cli/mod.rs
@@ -2,4 +2,5 @@
 
 pub mod components;
 pub mod config;
+pub mod db_type;
 pub mod ext;
