+++ crates/node-core/src/node_config.rs
@@ -6,17 +6,16 @@ use crate::{
         get_secret_key, DatabaseArgs, DebugArgs, DevArgs, NetworkArgs, PayloadBuilderArgs,
         PruningArgs, RpcServerArgs, TxPoolArgs,
     },
+    cl_events::ConsensusLayerHealthEvents,
     cli::{
         components::RethNodeComponentsImpl,
         config::{RethRpcConfig, RethTransactionPoolConfig},
         db_type::{DatabaseBuilder, DatabaseInstance},
         ext::{RethCliExt, RethNodeCommandConfig},
     },
-    commands::{
-        debug_cmd::EngineApiStore,
-        node::{cl_events::ConsensusLayerHealthEvents, events},
-    },
     dirs::{ChainPath, DataDirPath, MaybePlatformPath},
+    engine_api_store::EngineApiStore,
+    events,
     init::init_genesis,
     prometheus_exporter,
     utils::{get_single_header, write_peers_to_file},
